moved tensor tiling and concatenation to utils
