+++ earth2studio/models/dx/tempest_extremes.py
@@ -18,95 +18,13 @@ from rich.table import Table
 from rich.traceback import Traceback
 
 from earth2studio.io import KVBackend
-from earth2studio.utils.coords import CoordSystem, map_coords, split_coords
-
-
-def tile_xx_to_yy(
-    xx: torch.Tensor, xx_coords: CoordSystem, yy: torch.Tensor, yy_coords: CoordSystem
-) -> tuple[torch.Tensor, CoordSystem]:
-    """Tile tensor xx to match the leading dimensions of tensor yy.
-
-    Parameters
-    ----------
-    xx : torch.Tensor
-        Source tensor to be tiled
-    xx_coords : CoordSystem
-        Coordinate system for xx tensor
-    yy : torch.Tensor
-        Target tensor whose shape determines tiling
-    yy_coords : CoordSystem
-        Coordinate system for yy tensor
-
-    Returns
-    -------
-    Tuple[torch.Tensor, CoordSystem]
-        Tuple containing the tiled tensor and updated coordinate system
-
-    Raises
-    ------
-    ValueError
-        If xx has more dimensions than yy
-    """
-    n_lead = len(yy.shape) - len(xx.shape)
-
-    if n_lead < 0:
-        raise ValueError("xx must have fewer dimensions than yy.")
-
-    out_shape = yy.shape[:n_lead] + tuple([-1 for _ in range(len(xx.shape))])
-    out_coords = copy.deepcopy(yy_coords)
-    for key, val in xx_coords.items():
-        out_coords[key] = val
-
-    return xx.expand(out_shape), out_coords
-
-
-def cat_coords(
-    xx: torch.Tensor,
-    cox: CoordSystem,
-    yy: torch.Tensor,
-    coy: CoordSystem,
-    dim: str = "variable",
-) -> tuple[torch.Tensor, CoordSystem]:
-    """
-    concatenate data along coordinate dimension.
-
-    Parameters
-    ----------
-    xx : torch.Tensor
-        First input tensor which to concatenate
-    cox : CoordSystem
-        Ordered dict representing coordinate system that describes xx
-    yy : torch.Tensor
-        Second input tensor which to concatenate
-    coy : CoordSystem
-        Ordered dict representing coordinate system that describes yy
-    dim : str
-        name of dimension along which to concatenate
-
-    Returns
-    -------
-    tuple[torch.Tensor, CoordSystem]
-        Tuple containing output tensor and coordinate OrderedDict from
-        concatenated data.
-    """
-
-    if dim not in cox:
-        raise ValueError(f"dim {dim} is not in coords: {list(cox)}.")
-    if dim not in coy:
-        raise ValueError(f"dim {dim} is not in coords: {list(coy)}.")
-
-    # fix difference in latitude
-    _cox = cox.copy()
-    _cox["lat"] = coy["lat"]
-    xx, cox = map_coords(xx, cox, _cox)
-
-    coords = cox.copy()
-    dim_index = list(coords).index(dim)
-
-    zz = torch.cat((xx, yy), dim=dim_index)
-    coords[dim] = np.append(cox[dim], coy[dim])
-
-    return zz, coords
+from earth2studio.utils.coords import (
+    CoordSystem,
+    cat_coords,
+    map_coords,
+    split_coords,
+    tile_xx_to_yy,
+)
 
 
 class TempestExtremes:
