+++ recipes/hens/src/hens_ensemble.py
@@ -33,12 +33,16 @@ from earth2studio.io import IOBackend
 from earth2studio.models.dx import DiagnosticModel
 from earth2studio.models.px import PrognosticModel
 from earth2studio.perturbation import Perturbation
-from earth2studio.utils.coords import CoordSystem, map_coords, split_coords
+from earth2studio.utils.coords import (
+    CoordSystem,
+    cat_coords,
+    map_coords,
+    split_coords,
+)
 from earth2studio.utils.time import to_time_array
 
 from .hens_utilities import (
     TCTracking,
-    cat_coords,
     get_batchid_from_ensid,
     save_corrdiff_output,
 )
