+++ recipes/hens/src/hens_utilities.py
@@ -34,7 +34,6 @@ from earth2studio.models.auto import Package
 from earth2studio.models.dx import CorrDiff
 from earth2studio.models.px import PrognosticModel
 from earth2studio.perturbation import Perturbation
-from earth2studio.utils.coords import CoordSystem, map_coords
 from earth2studio.utils.time import to_time_array
 
 from .hens_utilities_reproduce import (
@@ -1022,55 +1021,6 @@ def get_batchid_from_ensid(
     )
 
     return batch_id
-
-
-def cat_coords(
-    xx: torch.Tensor,
-    cox: CoordSystem,
-    yy: torch.Tensor,
-    coy: CoordSystem,
-    dim: str = "variable",
-) -> tuple[torch.Tensor, CoordSystem]:
-    """
-    concatenate data along coordinate dimension.
-
-    Parameters
-    ----------
-    xx : torch.Tensor
-        First input tensor which to concatenate
-    cox : CoordSystem
-        Ordered dict representing coordinate system that describes xx
-    yy : torch.Tensor
-        Second input tensor which to concatenate
-    coy : CoordSystem
-        Ordered dict representing coordinate system that describes yy
-    dim : str
-        name of dimension along which to concatenate
-
-    Returns
-    -------
-    tuple[torch.Tensor, CoordSystem]
-        Tuple containing output tensor and coordinate OrderedDict from
-        concatenated data.
-    """
-
-    if dim not in cox:
-        raise ValueError(f"dim {dim} is not in coords: {list(cox)}.")
-    if dim not in coy:
-        raise ValueError(f"dim {dim} is not in coords: {list(coy)}.")
-
-    # fix difference in latitude
-    _cox = cox.copy()
-    _cox["lat"] = coy["lat"]
-    xx, cox = map_coords(xx, cox, _cox)
-
-    coords = cox.copy()
-    dim_index = list(coords).index(dim)
-
-    zz = torch.cat((xx, yy), dim=dim_index)
-    coords[dim] = np.append(cox[dim], coy[dim])
-
-    return zz, coords
 
 
 def save_corrdiff_output(cd_output_dict: dict, save_path: str) -> None:
