+++ test/models/dx/test_tempest_extremes.py
@@ -508,82 +508,7 @@ def test_async_record_state_waits_for_dump(mock_check):
 
 
 # ============================================================================
-# 8. UTILITY FUNCTION TESTS
-# ============================================================================
-
-
-def test_tile_xx_to_yy():
-    """Test tiling function for expanding dimensions"""
-    from collections import OrderedDict
-
-    from earth2studio.models.dx.tempest_extremes import tile_xx_to_yy
-
-    xx = torch.randn(2, 721, 1440)
-    xx_coords = OrderedDict(
-        {
-            "variable": np.array(["z", "lsm"]),
-            "lat": np.linspace(90, -90, 721),
-            "lon": np.linspace(0, 360, 1440),
-        }
-    )
-
-    yy = torch.randn(3, 4, 5, 721, 1440)
-    yy_coords = OrderedDict(
-        {
-            "ensemble": np.array([0, 1, 2]),
-            "time": np.array([1, 2, 3, 4]),
-            "lead_time": np.array([0, 1, 2, 3, 4]),
-            "lat": np.linspace(90, -90, 721),
-            "lon": np.linspace(0, 360, 1440),
-        }
-    )
-
-    result, result_coords = tile_xx_to_yy(xx, xx_coords, yy, yy_coords)
-
-    # Result should have yy's leading dims + all of xx's dims
-    # yy.shape = (3, 4, 5, 721, 1440), xx.shape = (2, 721, 1440)
-    # n_lead = 5 - 3 = 2, so we prepend yy's first 2 dims to xx
-    # Result shape should be (3, 4, 2, 721, 1440)
-    assert result.shape == (3, 4, 2, 721, 1440)
-    assert "variable" in result_coords
-    assert "time" in result_coords
-
-
-def test_cat_coords():
-    """Test coordinate concatenation"""
-    from collections import OrderedDict
-
-    from earth2studio.models.dx.tempest_extremes import cat_coords
-
-    xx = torch.randn(1, 2, 721, 1440)
-    cox = OrderedDict(
-        {
-            "time": np.array([0]),
-            "variable": np.array(["u10m", "v10m"]),
-            "lat": np.linspace(90, -90, 721),
-            "lon": np.linspace(0, 360, 1440),
-        }
-    )
-
-    yy = torch.randn(1, 1, 721, 1440)
-    coy = OrderedDict(
-        {
-            "time": np.array([0]),
-            "variable": np.array(["msl"]),
-            "lat": np.linspace(90, -90, 721),
-            "lon": np.linspace(0, 360, 1440),
-        }
-    )
-
-    result, result_coords = cat_coords(xx, cox, yy, coy, dim="variable")
-
-    assert result.shape == (1, 3, 721, 1440)
-    assert len(result_coords["variable"]) == 3
-    assert np.array_equal(result_coords["variable"], ["u10m", "v10m", "msl"])
-
-
-# ============================================================================
-# 9. ERROR HANDLING TESTS
+# 8. ERROR HANDLING TESTS
 # ============================================================================
 
 
@@ -647,7 +572,7 @@ def test_async_check_for_failures_raises(mock_check):
 
 
 # ============================================================================
-# 10. CLEANUP TESTS
+# 9. CLEANUP TESTS
 # ============================================================================
